+++ packages/nvidia_nat_core/tests/nat/server/test_unified_api_server.py
@@ -346,7 +346,7 @@ async def test_generate_endpoint(client: httpx.AsyncClient, config: Config):
 
 async def test_generate_endpoint_returns_error_body_when_workflow_raises(client: httpx.AsyncClient, config: Config):
     """When the workflow raises, non-streaming generate returns 422 with Error JSON body."""
-    with patch("nat.front_ends.fastapi.fastapi_front_end_plugin_worker.generate_single_response") as mock_gen:
+    with patch("nat.front_ends.fastapi.routes.common_utils.generate_single_response") as mock_gen:
         mock_gen.side_effect = NotImplementedError("No human prompt callback was registered.")
         input_message = {"message": "hello"}
         response = await client.post(f"{config.endpoint.generate}", json=input_message)
@@ -394,7 +394,7 @@ async def test_chat_endpoint(client: httpx.AsyncClient, config: Config):
 
 async def test_chat_endpoint_returns_error_body_when_workflow_raises(client: httpx.AsyncClient, config: Config):
     """When the workflow raises, non-streaming chat returns 422 with Error JSON body."""
-    with patch("nat.front_ends.fastapi.fastapi_front_end_plugin_worker.generate_single_response") as mock_gen:
+    with patch("nat.front_ends.fastapi.routes.common_utils.generate_single_response") as mock_gen:
         mock_gen.side_effect = NotImplementedError("No human prompt callback was registered.")
         input_message = {"messages": [{"role": "user", "content": "hello"}], "use_knowledge_base": True}
         response = await client.post(f"{config.endpoint.chat}", json=input_message)
